client/requirements.txt
pygame
windows-curses
python-socketio
websocket-client    # python-socketio requires this